Minor: Edits around existing claims.
I'd like to make the current claims feel "more or less" true, up to the
level of uncertainty in "Han shot first," but generally what I would
personally agree with in casual conversation. I'd also like to aim for
highlighted tidbits (as in definitions, claims, etc) that serve as a
nice outline to the overall article when you skim them.