Avoid using both falsity and falseness.
I don't particularly love either word. Google Trends indicates that
"falsity" is more common, but in the one place I'm using either word
now, I just feel like "falseness" is a better fit for the context.

So I'm going with: rare use of falseness is ok, but falsity is not ok.

For now.